move sentence, word
