+++ packages/cargo/jns42-core/src/lib.rs
@@ -1,4 +1,5 @@
 mod ffi;
 pub mod models;
+pub mod names;
 pub mod schema_transforms;
 pub mod utils;
