+++ packages/cargo/jns42-core/src/names/mod.rs
@@ -1,0 +1,5 @@
+mod sentence;
+mod word;
+
+pub use sentence::*;
+pub use word::*;
