+++ packages/cargo/jns42-core/src/names/sentence.rs
@@ -1,36 +1,8 @@
+use super::Word;
 use std::{
   iter::{empty, once},
   slice::Iter,
 };
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Word(String);
-
-impl Word {
-  fn new(value: &str) -> Self {
-    assert!(!value.is_empty());
-
-    Self(value.to_lowercase())
-  }
-
-  pub fn as_lower(&self) -> &str {
-    &self.0
-  }
-
-  pub fn to_upper(&self) -> String {
-    self.0.to_uppercase()
-  }
-
-  pub fn to_pascal(&self) -> String {
-    self.0[0..1].to_uppercase() + &self.0[1..]
-  }
-}
-
-impl AsRef<str> for Word {
-  fn as_ref(&self) -> &str {
-    &self.0
-  }
-}
 
 #[derive(Debug, Default, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Sentence(Vec<Word>);
