+++ packages/cargo/jns42-core/src/utils/mod.rs
@@ -2,5 +2,4 @@ pub mod json_pointer;
 pub mod merge;
 pub mod names;
 pub mod product;
-pub mod sentence;
 pub mod url;
